registration page
added form inputs, added validation
