+++ HomeHospital-Frontend/hh-front/src/App.js
@@ -1,14 +1,24 @@
 import "bootstrap/dist/css/bootstrap.min.css";
+import React, { useEffect } from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import Home from "./pages/Home";
 import Login from "./pages/Login";
 import Register from "./pages/Register";
 
 function App() {
+  useEffect(() => {
+    document.title = "HomeHospital";
+  }, []);
+
   return (
     <Router>
       <Routes>
-        <Route path="/" element={<Login className="full-height" />} />
-        <Route path="/register" element={<Register />} />
+        <Route path="/" element={<Home />} />
+        <Route path="/login" element={<Login />} />
+        <Route
+          path="/register"
+          element={<Register className="full-height" />}
+        />
       </Routes>
     </Router>
   );
