+++ HomeHospital-Frontend/hh-front/src/pages/Home.js
@@ -1,0 +1,28 @@
+import React from "react";
+import { Button } from "react-bootstrap";
+import { useNavigate } from "react-router-dom";
+
+function Home() {
+  let navigate = useNavigate();
+  return (
+    <div>
+      <h1>Home Page</h1>
+      <Button
+        onClick={() => {
+          navigate("/login");
+        }}
+      >
+        Login
+      </Button>
+      <Button
+        onClick={() => {
+          navigate("/register");
+        }}
+      >
+        Register
+      </Button>
+    </div>
+  );
+}
+
+export default Home;
