--- HomeHospital-Frontend/hh-front/src/pages/Login.module.css
@@ -1,28 +1,0 @@
-.boxWidth {
-  width: 50%;
-  margin: auto;
-  background-color: white;
-  border-radius: 0px 50px 0px 30px;
-}
-
-.leftSide {
-  position: relative;
-  background-color: #67d8c6;
-  box-shadow: -5px 10px 5px 2px #5f5f5f;
-  border-radius: 0px 0px 0px 50px;
-  padding: 0%;
-}
-
-.rightSide {
-  background-color: #394b63;
-  box-shadow: 5px 10px 5px 2px #5f5f5f;
-  border-radius: 0px 50px 0px 0px;
-}
-
-.title {
-  position: absolute;
-  top: 60px;
-  left: 125px;
-  color: white;
-  text-align: center;
-}
